🧬 Rename the module as *config* to be more clear.
