+++ test/unit_test/command/process.py
@@ -12,7 +12,7 @@ import pytest
 from yaml import load as yaml_load
 
 from pymock_api.model.openapi._parse import OpenAPIParser
-from pymock_api.model.openapi.swagger_config import set_component_definition
+from pymock_api.model.openapi.config import set_component_definition
 
 try:
     from yaml import CLoader as Loader
